Create platform/overview package and move classes
