+++ src/main/java/io/cryostat/MainModule.java
@@ -51,13 +51,13 @@ import io.cryostat.core.log.Logger;
 import io.cryostat.core.sys.Environment;
 import io.cryostat.core.tui.ClientWriter;
 import io.cryostat.messaging.MessagingModule;
-import io.cryostat.net.AbstractNode.BaseNodeType;
 import io.cryostat.net.NetworkModule;
 import io.cryostat.net.web.http.HttpMimeType;
 import io.cryostat.platform.PlatformModule;
 import io.cryostat.platform.internal.CustomTargetPlatformClient.CustomTargetNodeType;
 import io.cryostat.platform.internal.DefaultPlatformClient.JDPNodeType;
 import io.cryostat.platform.internal.KubeApiPlatformClient.KubernetesNodeType;
+import io.cryostat.platform.overview.BaseNodeType;
 import io.cryostat.recordings.RecordingsModule;
 import io.cryostat.rules.Rule;
 import io.cryostat.rules.RulesModule;
