+++ src/main/java/io/cryostat/platform/PlatformClient.java
@@ -41,7 +41,7 @@ import java.io.IOException;
 import java.util.List;
 import java.util.function.Consumer;
 
-import io.cryostat.net.EnvironmentNode;
+import io.cryostat.platform.overview.EnvironmentNode;
 
 public interface PlatformClient {
     void start() throws IOException;
