+++ src/main/java/io/cryostat/platform/internal/CustomTargetPlatformClient.java
@@ -53,11 +53,11 @@ import javax.inject.Named;
 import io.cryostat.MainModule;
 import io.cryostat.core.net.discovery.JvmDiscoveryClient.EventKind;
 import io.cryostat.core.sys.FileSystem;
-import io.cryostat.net.AbstractNode.BaseNodeType;
-import io.cryostat.net.AbstractNode.NodeType;
-import io.cryostat.net.EnvironmentNode;
-import io.cryostat.net.TargetNode;
 import io.cryostat.platform.ServiceRef;
+import io.cryostat.platform.overview.BaseNodeType;
+import io.cryostat.platform.overview.EnvironmentNode;
+import io.cryostat.platform.overview.NodeType;
+import io.cryostat.platform.overview.TargetNode;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
