+++ src/main/java/io/cryostat/platform/internal/KubeApiPlatformClient.java
@@ -50,12 +50,12 @@ import java.util.stream.Collectors;
 import io.cryostat.core.log.Logger;
 import io.cryostat.core.net.JFRConnectionToolkit;
 import io.cryostat.core.net.discovery.JvmDiscoveryClient.EventKind;
-import io.cryostat.net.AbstractNode.BaseNodeType;
-import io.cryostat.net.AbstractNode.NodeType;
-import io.cryostat.net.EnvironmentNode;
-import io.cryostat.net.TargetNode;
 import io.cryostat.platform.ServiceRef;
 import io.cryostat.platform.ServiceRef.AnnotationKey;
+import io.cryostat.platform.overview.BaseNodeType;
+import io.cryostat.platform.overview.EnvironmentNode;
+import io.cryostat.platform.overview.NodeType;
+import io.cryostat.platform.overview.TargetNode;
 import io.cryostat.util.URIUtil;
 
 import dagger.Lazy;
