+++ src/main/java/io/cryostat/platform/internal/KubeEnvPlatformClient.java
@@ -48,8 +48,8 @@ import java.util.stream.Collectors;
 import io.cryostat.core.log.Logger;
 import io.cryostat.core.net.JFRConnectionToolkit;
 import io.cryostat.core.sys.Environment;
-import io.cryostat.net.EnvironmentNode;
 import io.cryostat.platform.ServiceRef;
+import io.cryostat.platform.overview.EnvironmentNode;
 import io.cryostat.util.URIUtil;
 
 import dagger.Lazy;
